+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Use `<PersistedPaneset>` smart component for Data Import landing page (UIDATIMP-884)
 * Cover `<LogViewer>` component with unit tests (UIDATIMP-710)
 * Cover `<FileItem>` component with unit tests (UIDATIMP-716)
+* Cover `<ProfileTree>` component with unit tests (UIDATIMP-718)
 
 ### Bugs fixed:
 * Job profile fails because 1 action profile has 2 field mappings attached to it (UIDATIMP-931)